Optimize memory and footprint of icu NameIndexToNumericData
Shrink from 30kb to 13.5kb and use ReadOnlySpan<byte> to avoid initalizing a large ushort[] in IcuLocaleData.cctor
